Patching?! And cards
Created a Anxious Mind, a use it now or give something up card. This has been added as a flip for Creative Mind.

Creative Thoughts and Racing Thoughts now gets ethereal and exhaust on created cards

Removed pregenerated cards from pools and compendium

Created Dream and Plan
Dream is like Limit Break for strength and Dex but more expensive and is a flip of Plan which is a cheap draw.

Fixed ExploitVirus not showing or swapping to upgraded ExploitPhish

Found a bug that was causing Hyper Focus to affect enemies too (Very noticeable when using Dark Shackles on them) and added a failsafe to prevent

Reduced damage on Hyperkick to be in line with Twin strike

Created new starter card to take the place of Concentration - Restore Point

Restore Point is a random return from exhaust card that upgrades from 2 cost to 1 cost

Fixed Lasers not scaling with Dexterity in text / damage

Changed energy gain on ReadyWeapons to Energize instead

Changed Character Background to updated picture, added patch from Mistress Autumn that makes the text much easier to read on character select screen

Updated some cardstrings, relicstrings, and powerstrings to properly reflect mechanics.